# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.ConfirmPaymentRequest:
    properties:
      booking_id:
        example: 1024
        type: integer
      transaction_id:
        example: 5001
        type: integer
    required:
    - booking_id
    - transaction_id
    type: object
...
  dto.PaymentMethodDTO:
    properties:
      id:
        example: 2
        type: integer
      logo:
        example: /img/pm/DANA.png
        type: string
      name:
        example: DANA
        type: string
    type: object
  dto.PaymentPageResponse:
    properties:
      booking_id:
        example: 1024
        type: integer
      cinema_name:
        example: CGV Grand Indonesia
        type: string
      email:
        example: ali.mustadji@example.com
        type: string
      full_name:
        example: Ali Mustadji
        type: string
      movie_title:
        example: Interstellar Reborn
        type: string
      payment_methods:
        items:
          $ref: '#/definitions/dto.PaymentMethodDTO'
        type: array
      phone:
        example: "081234567890"
        type: string
      quantity:
        example: 2
        type: integer
      seat_labels:
        example: C7, C8
        type: string
      show_date:
        example: "2026-06-15T00:00:00Z"
        type: string
      show_time:
        example: "18:30"
        type: string
      total_price:
        example: 100000
        type: integer
    type: object
...
  dto.ResponseError:
    properties:
      error:
        example: internal server error/bad request
        type: string
      message:
        example: Failed get data/internal server error
        type: string
      status:
        example: error
        type: string
    type: object
...
  dto.SubmitPaymentRequest:
    properties:
      booking_id:
        example: 56
        type: integer
      payment_method_id:
        example: 3
        type: integer
    required:
    - booking_id
    - payment_method_id
    type: object
...
  dto.TicketResultResponse:
    properties:
      category:
        example: 13+
        type: string
      movie_title:
        example: Interstellar Reborn
        type: string
      payment_status:
        example: PAID
        type: string
      qr_code:
        example: TICKITZ-TX-5001-SECURE-STRING
        type: string
      seat_labels:
        example: C7, C8
        type: string
      show_date:
        example: "2026-06-15T00:00:00Z"
        type: string
      show_time:
        example: "18:30"
        type: string
      ticket_count:
        example: 2
        type: integer
      total_price:
        example: 100000
        type: integer
    type: object
...
  dto.TransactionModalResponse:
    properties:
      payment_deadline:
        example: "2026-06-13T19:47:00Z"
        type: string
      status:
        example: PENDING
        type: string
      total_price:
        example: 100000
        type: integer
      transaction_id:
        example: 5001
        type: integer
      virtual_rek:
        example: 8856001234567890
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: limustadji@gmail.com
    name: alpha male
    url: https://github.com/L1mus
  description: This is a backend project for a web application called E-wallet.
  license:
    name: MIT
  title: Movie API
  version: "1.0"
paths:
...
  /transactions/checkout:
    post:
      consumes:
      - application/json
      description: Triggers a direct backend-to-backend API request to DOKU Core Sandbox
        System to simulate a payment invoice creation.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully received simulated checkout payload from DOKU
            Core
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error - Failed communicating with external
            payment gateway API
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Simulate DOKU Sandbox Checkout Invoice
      tags:
      - DOKU Integration
  /transactions/confirm:
    post:
      consumes:
      - application/json
      description: Verify and process transaction clearance (simulates payment settlement).
      parameters:
      - description: Payload containing transaction ID and booking ID to be confirmed
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ConfirmPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment settlement verified and confirmed successfully
          schema:
            $ref: '#/definitions/dto.TicketResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "406":
          description: Not Acceptable - The transaction has already been paid
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Simulate/Confirm Payment Complete
      tags:
      - Transactions
  /transactions/doku-callback:
    post:
      consumes:
      - application/json
      description: Automated HTTP callback listener endpoint designed to receive status
        notification updates directly pushed by DOKU gateway systems.
      parameters:
      - description: Raw payload notification structure sent from DOKU
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Returns status verification token acknowledgement back to DOKU
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request - Invalid payload structure
          schema:
            additionalProperties: true
            type: object
      summary: DOKU Webhook Notification Handler Listener
      tags:
      - DOKU Integration
  /transactions/payment/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve checkout summary details (ticket quantity, seats, total
        price, user profile) and available payment options.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved payment information summary
          schema:
            $ref: '#/definitions/dto.PaymentPageResponse'
        "401":
          description: Unauthorized - Login session expired or token invalid
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "404":
          description: Not Found - Booking record not found or does not belong to
            user
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get Payment Page Information
      tags:
      - Transactions
  /transactions/qr/{id}:
    get:
      description: Generates and serves a raw binary PNG QR code image linked to the
        transaction for entry scanning verification.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/png
      responses:
        "200":
          description: Returns raw binary PNG image of the QR Code
          schema:
            type: file
        "500":
          description: Internal Server Error - Failed to generate QR code image output
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Generate Ticket QR Code Image
      tags:
      - Transactions
  /transactions/result/{id}:
    get:
      consumes:
      - application/json
      description: Fetch the finalized digital ticket receipt details after a successful
        transaction payment.
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved final digital ticket information
          schema:
            $ref: '#/definitions/dto.TicketResultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get Digital Ticket Invoice
      tags:
      - Transactions
  /transactions/submit:
    post:
      consumes:
      - application/json
      description: Process the chosen payment method to initiate a transaction and
        generate payment details (e.g., Virtual Account).
      parameters:
      - description: Payload containing booking ID and selected payment method ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment method submitted successfully
          schema:
            $ref: '#/definitions/dto.TransactionModalResponse'
        "400":
          description: Bad Request - Missing or invalid JSON body parameters
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "403":
          description: Forbidden - Access denied, this booking does not belong to
            the authenticated user
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Submit Selected Payment Method
      tags:
      - Transactions
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token used for authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"